// src/store/main/analysis/type.ts
/*
 * @Author: your name
 * @Date: 2021-09-12 11:07:27
 * @LastEditTime: 2021-09-12 11:10:26
 * @LastEditors: Please set LastEditors
 * @Description: In User Settings Edit
 * @FilePath: /vue3-practice/src/store/main/analysis/tyoe.ts
 */
export interface IDashboardState {
  categoryGoodsCount: any[]
  categoryGoodsSale: any[]
  categoryGoodsFavor: any[]
  addressGoodsSale: any[]
}